requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.39.1
django-cors-headers==3.1.1
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.1.9
django-redis==4.10.0
django-rest-auth==0.9.5
django-telegram-login==0.2.3
django-webpack-loader==0.6.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
dnspython==1.16.0
email-validator==1.0.4
idna==2.8
mysqlclient==1.4.4
oauthlib==3.0.2
Pillow==6.2.1
psycopg2==2.8.4
py3dns==3.2.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==2.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.6
